+++ src/index.d.ts
@@ -276,6 +276,11 @@ declare class RxChangeEvent {
     toJSON(): any;
 }
 
+declare class RxError extends Error {
+    rxdb: boolean; // always true, use this to detect if its an rxdb-error
+    parameters: any; // an object with parameters to use the programatically
+}
+
 export interface DatabaseCreator {
     name: string;
     adapter: any;
@@ -310,6 +315,7 @@ export {
     RxChangeEvent as RxChangeEvent,
     PouchDB as PouchDB,
     RxCollectionCreator as RxCollectionCreator,
+    RxError as RxError
 };
 
 export default {
@@ -324,5 +330,6 @@ export default {
     RxSchema,
     PouchDB,
     QueryChangeDetector,
-    RxDatabase
+    RxDatabase,
+    RxError
 };